Move watching for confirmations into a shared saga
Generalizes the code a bit so that we don't have to duplicate
the same functionality in both the redemption and deposit flows
